Improve showdown extension handling
- Adds a new 'extract' param that allows extensions to provide details
of the items that should be extracted before they run.

- This is moved into showdown itself to prevent extensive duplication
and missing extraction handlers in the extensions themselves.

- The result should be more reliable application of extensions.

- Each part of GhostGFM exists both inside of ghostgfm and as a
standalone extension. This aids testing each one as well as testing
them in combination. In Ghost we can choose to use one or the other.

- All of the extensions have been updated.

- A vast array of new tests have been added, including ones from
issues raised against Ghost + the style test post from demo.ghost.io